+++ lib/spout/helpers/array_statistics.rb
@@ -1,41 +1,46 @@
+# Extensions to the Array class to calculate quartiles, outliers, and statistics
 class Array
+  def compact_empty
+    compact.reject { |a| a.is_a?(Spout::Models::Empty) }
+  end
+
   def n
-    self.compact.count
+    compact_empty.count
   end
 
   def mean
-    array = self.compact
+    array = compact_empty
     return nil if array.size == 0
     array.inject(:+).to_f / array.size
   end
 
   def sample_variance
-    array = self.compact
+    array = compact_empty
     m = array.mean
-    sum = array.inject(0){|accum, i| accum +(i-m)**2 }
+    sum = array.inject(0) { |a, e| a + (e - m)**2 }
     sum / (array.length - 1).to_f
   end
 
   def standard_deviation
-    array = self.compact
+    array = compact_empty
     return nil if array.size < 2
-    return Math.sqrt(array.sample_variance)
+    Math.sqrt(array.sample_variance)
   end
 
   def median
-    array = self.compact.sort
+    array = compact_empty.sort
     return nil if array.size == 0
     len = array.size
-    len % 2 == 1 ? array[len/2] : (array[len/2 - 1] + array[len/2]).to_f / 2
+    len.odd? ? array[len / 2] : (array[len / 2 - 1] + array[len / 2]).to_f / 2
   end
 
   def unknown
-    self.select{|s| s == nil}.count
+    count { |a| a.is_a?(Spout::Models::Empty) }
   end
 
   def quartile_sizes
-    quartile_size = self.count / 4
-    quartile_fraction = self.count % 4
+    quartile_size = count / 4
+    quartile_fraction = count % 4
 
     quartile_sizes = [quartile_size] * 4
     (0..quartile_fraction - 1).to_a.each do |index|
@@ -46,75 +51,76 @@ class Array
   end
 
   def quartile_one
-    self[0..(self.quartile_sizes[0] - 1)]
+    self[0..(quartile_sizes[0] - 1)]
   end
 
   def quartile_two
-    sizes = self.quartile_sizes
+    sizes = quartile_sizes
     start = sizes[0]
     stop = start + sizes[1] - 1
     self[start..stop]
   end
 
   def quartile_three
-    sizes = self.quartile_sizes
+    sizes = quartile_sizes
     start = sizes[0] + sizes[1]
     stop = start + sizes[2] - 1
     self[start..stop]
   end
 
   def quartile_four
-    sizes = self.quartile_sizes
+    sizes = quartile_sizes
     start = sizes[0] + sizes[1] + sizes[2]
     stop = start + sizes[3] - 1
     self[start..stop]
   end
 
   def compact_min
-    self.compact.min
+    compact_empty.min
   end
 
   def compact_max
-    self.compact.max
+    compact_empty.max
   end
 
   def outliers
-    array = self.compact.sort.select{|v| v.is_a?(Numeric)}
+    array = compact_empty.sort.select { |v| v.is_a?(Numeric) }
     q1 = (array.quartile_one + array.quartile_two).median
     q3 = (array.quartile_three + array.quartile_four).median
-    return [] if q1 == nil or q3 == nil
+    return [] if q1.nil? || q3.nil?
     iq_range = q3 - q1
     inner_fence_lower = q1 - iq_range * 1.5
     inner_fence_upper = q3 + iq_range * 1.5
-    outer_fence_lower = q1 - iq_range * 3
-    outer_fence_upper = q3 + iq_range * 3
-    array.select{ |v| v > inner_fence_upper or v < inner_fence_lower }
+    array.select { |v| v > inner_fence_upper || v < inner_fence_lower }
   end
 
   def major_outliers
-    array = self.compact.sort.select{|v| v.is_a?(Numeric)}
+    array = compact_empty.sort.select { |v| v.is_a?(Numeric) }
     q1 = (array.quartile_one + array.quartile_two).median
     q3 = (array.quartile_three + array.quartile_four).median
-    return [] if q1 == nil or q3 == nil
+    return [] if q1.nil? || q3.nil?
     iq_range = q3 - q1
-    inner_fence_lower = q1 - iq_range * 1.5
-    inner_fence_upper = q3 + iq_range * 1.5
     outer_fence_lower = q1 - iq_range * 3
     outer_fence_upper = q3 + iq_range * 3
-    array.select{ |v| v > outer_fence_upper or v < outer_fence_lower }
+    array.select { |v| v > outer_fence_upper || v < outer_fence_lower }
   end
 
   def minor_outliers
-    self.outliers - self.major_outliers
+    outliers - major_outliers
   end
-
 end
 
 module Spout
   module Helpers
     class ArrayStatistics
       def self.calculations
-        [["N", :n, :count], ["Mean", :mean, :decimal], ["StdDev", :standard_deviation, :decimal, "± %s"], ["Median", :median, :decimal], ["Min", :compact_min, :decimal], ["Max", :compact_max, :decimal], ["Unknown", :unknown, :count]]
+        [['N', :n, :count],
+         ['Mean', :mean, :decimal],
+         ['StdDev', :standard_deviation, :decimal, '± %s'],
+         ['Median', :median, :decimal],
+         ['Min', :compact_min, :decimal],
+         ['Max', :compact_max, :decimal],
+         ['Unknown', :unknown, :count]]
       end
     end
   end