+++ lib/spout/models/domain.rb
@@ -5,9 +5,7 @@ require 'spout/models/option'
 
 module Spout
   module Models
-
     class Domain < Spout::Models::Record
-
       attr_accessor :id, :folder, :options
       attr_reader :errors
 
@@ -26,11 +24,11 @@ module Spout
             nil
           end
         rescue => e
-          @errors << "Parsing error found in #{@id}.json: #{e.message}" if file_name != nil
+          @errors << "Parsing error found in #{@id}.json: #{e.message}" unless file_name.nil?
           nil
         end
 
-        if json and json.is_a? Array
+        if json && json.is_a?(Array)
           @id = file_name.to_s.gsub(/^(.*)\/|\.json$/, '').downcase
           @options = (json || []).collect do |option|
             Spout::Models::Option.new(option)
@@ -38,10 +36,7 @@ module Spout
         elsif json
           @errors << "Domain must be a valid array in the following format: [\n  {\n    \"value\": \"1\",\n    \"display_name\": \"First Choice\",\n    \"description\": \"First Description\"\n  },\n  {\n    \"value\": \"2\",\n    \"display_name\": \"Second Choice\",\n    \"description\": \"Second Description\"\n  }\n]"
         end
-
       end
-
-
     end
   end
 end