+++ lib/spout/models/graphables/histogram.rb
@@ -3,8 +3,8 @@ require 'spout/models/graphables/default'
 module Spout
   module Models
     module Graphables
+      # Generates data for variable histograms
       class Histogram < Spout::Models::Graphables::Default
-
         def title
           @variable.display_name
         end
@@ -23,18 +23,18 @@ module Spout
 
         def series
           @chart_variable.domain.options.collect do |option|
-            visit_subjects = @subjects.select{ |s| s.send(@chart_variable.id) == option.value and s.send(@variable.id) != nil } rescue visit_subjects = []
-            visit_subject_values = visit_subjects.collect(&@variable.id.to_sym).sort rescue visit_subject_values = []
+            visit_subjects = @subjects.select{ |s| s.send(@chart_variable.id) == option.value && !s.send(@variable.id).nil? && !s.send(@variable.id).is_a?(Spout::Models::Empty) } rescue visit_subjects = []
+            visit_subject_values = visit_subjects.collect(&@variable.id.to_sym).sort # rescue visit_subject_values = []
             next unless visit_subject_values.size > 0
 
             data = []
 
             if @variable.type == 'choices'
-              data = filtered_domain_options(@variable).collect do |option|
-                visit_subject_values.select{ |v| v == option.value }.count
+              data = filtered_domain_options(@variable).collect do |o|
+                visit_subject_values.select { |v| v == o.value }.count
               end
             else
-              visit_subject_values.group_by{|v| get_bucket(v) }.each do |key, values|
+              visit_subject_values.group_by { |v| get_bucket(v) }.each do |key, values|
                 data[categories.index(key)] = values.count if categories.index(key)
               end
             end
@@ -46,7 +46,6 @@ module Spout
         def x_axis_title
           @variable.units
         end
-
       end
     end
   end