+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 ## 0.11.0
 
 ### Enhancements
+- **Deploy Command**
+  - The deploy command can now accept flags and arguments that modify and scope graphs and images
+    - ex: `spout deploy WEBSERVER var1 var2 --rows=100 --clean`
+    - This pushes a clean version of graphs and images based on the first 100 rows in the CSVs for var1 and var2
+- **Graph Command**
+  - Tables no longer include unknown values for variables that don't exist across other CSVs with the same subject and visit
+  - Tables still track unknown values if the CSV contains a column for the variable, but the cell for subject is blank
+  - Histograms of variables that only contain a small range (less than 12) of integers now use discrete buckets of size 1 instead of creating fractional buckets
+- **Image Command**
+  - The image command now correctly filters specified variables
+    - ex: `spout p var1` would only generate an image for `var1` variable
 - **Gem Changes**
   - Use of Ruby 2.2.3 is now recommended